refactor: migrate Jest setup to jest.ts and remove deprecated setup files
